PERF: Unnecessary modules were removed and DTILesionTrack python was improved
The python module (DTILesionTrack) was reformulated in all parts that calls the slicer.util.loadVolume, where it address directly the Node to a variable.
This strategy prevents errors when the module is running more than once.
The unnecessary modules (BrainExtractionTool and BrainTissueMask) were removed from its extension because these modules now are in separeted projects (BrainTissuesExtensions and ROBEXBrainExtraction)